fix: android config worked out for local dev and react-native-image-colors patch refactor: remove cycle dep from WhoSampled.service
