+++ android/app/src/main/java/com/bedfellow/MainActivity.kt
@@ -1,3 +1,5 @@
+package com.bedfellow
+
 import com.facebook.react.ReactActivity
 import com.facebook.react.ReactActivityDelegate
 import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint.fabricEnabled
