+++ android/app/src/main/java/com/bedfellow/MainApplication.kt
@@ -1,4 +1,4 @@
-package com.bedfellow;
+package com.bedfellow
 
 import android.app.Application
 import com.facebook.react.PackageList
