+++ src/services/whosampled/WhoSampled.service.ts
@@ -98,26 +98,3 @@ const getWhoSampledDocument = async (urlFragment: string, variant: CONNECTIONS):
   }
   return null;
 };
-
-export const getWhoSampledImage = async (url: string | null): Promise<string | null> => {
-  if (!url) {
-    return null;
-  }
-
-  try {
-    const result = await axios.get(url, {
-      responseType: 'blob',
-    });
-    return await blobToBase64(result.data);
-  } catch (err) {
-    return null;
-  }
-};
-
-const blobToBase64 = (blob: Blob): Promise<string> => {
-  return new Promise((resolve, _) => {
-    const reader = new FileReader();
-    reader.onloadend = () => resolve(reader.result as string);
-    reader.readAsDataURL(blob);
-  });
-};
