+++ src/services/whosampled/utilities/utilities.ts
@@ -1,7 +1,7 @@
 import { load } from 'cheerio';
+import axios from 'axios';
 import { Sample } from '../../../types/whosampled';
 import { HEADER_TITLES } from '../enums';
-import { getWhoSampledImage } from '../WhoSampled.service';
 
 const WHOSAMPLED_BASE_URL = 'https://www.whosampled.com';
 
@@ -57,6 +57,29 @@ const parseWhoSampledPage = async (document: string, headerText: HEADER_TITLES):
     })
   );
   return samplesWithImages;
+};
+
+export const getWhoSampledImage = async (url: string | null): Promise<string | null> => {
+  if (!url) {
+    return null;
+  }
+
+  try {
+    const result = await axios.get(url, {
+      responseType: 'blob',
+    });
+    return await blobToBase64(result.data);
+  } catch (err) {
+    return null;
+  }
+};
+
+const blobToBase64 = (blob: Blob): Promise<string> => {
+  return new Promise((resolve, _) => {
+    const reader = new FileReader();
+    reader.onloadend = () => resolve(reader.result as string);
+    reader.readAsDataURL(blob);
+  });
 };
 
 export default parseWhoSampledPage;
